+++ tests/TestCase/TestSuite/TestTraitTest.php
@@ -189,8 +189,14 @@ class TestTraitTest extends TestCase
     {
         $files = [create_tmp_file(), create_tmp_file()];
         $this->assertFileExists($files[0]);
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         $this->assertFileExists($files);
         $this->assertFileExists(new ExampleOfTraversable($files));
+        error_reporting($errorReporting);
+
+        $this->expectException(Deprecated::class);
+        $this->assertFileExists($files);
     }
 
     /**
@@ -201,12 +207,18 @@ class TestTraitTest extends TestCase
     {
         $this->assertFileExtension('jpg', 'file.jpg');
         $this->assertFileExtension('jpeg', 'FILE.JPEG');
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         $this->assertFileExtension('jpg', [
             'file.jpg',
             'file.JPG',
             'path/to/file.jpg',
             '/full/path/to/file.jpg',
         ]);
+        error_reporting($errorReporting);
+
+        $this->expectException(Deprecated::class);
+        $this->assertFileExtension('jpg', ['file.jpg', 'file.JPG']);
     }
 
     /**
@@ -217,6 +229,12 @@ class TestTraitTest extends TestCase
     {
         $files = [create_tmp_file('string'), create_tmp_file('string')];
         $this->assertFileMime($files[0], 'text/plain');
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
+        $this->assertFileMime($files, 'text/plain');
+        error_reporting($errorReporting);
+
+        $this->expectException(Deprecated::class);
         $this->assertFileMime($files, 'text/plain');
     }
 
@@ -228,8 +246,14 @@ class TestTraitTest extends TestCase
     {
         $files = [TMP . 'noExisting1', TMP . 'noExisting2'];
         $this->assertFileNotExists($files[0]);
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         $this->assertFileNotExists($files);
         $this->assertFileNotExists(new ExampleOfTraversable($files));
+        error_reporting($errorReporting);
+
+        $this->expectException(Deprecated::class);
+        $this->assertFileNotExists($files);
     }
 
     /**
@@ -245,12 +269,18 @@ class TestTraitTest extends TestCase
         $this->assertFilePerms($files[0], ['0600', '0666']);
         $this->assertFilePerms($files[0], [0600, 0666]);
         $this->assertFilePerms($files[0], ['0600', 0666]);
+
+        $errorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
         $this->assertFilePerms($files, '0600');
         $this->assertFilePerms($files, 0600);
         $this->assertFilePerms($files, ['0600', '0666']);
         $this->assertFilePerms($files, [0600, 0666]);
         $this->assertFilePerms(new ExampleOfTraversable($files), '0600');
         $this->assertFilePerms(new ExampleOfTraversable($files), ['0600', '0666']);
+        error_reporting($errorReporting);
+
+        $this->expectException(Deprecated::class);
+        $this->assertFilePerms($files, '0600');
     }
 
     /**